Change blend shape mask array format enforcement in FBXDocument::_parse_meshes()
The code change updates the condition for enforcing the blend shape mask array format in the _parse_meshes() function of FBXDocument. It now uses static_cast<int64_t>(1) instead of just 1 to ensure compatibility with different platforms.

+++ fbx_document.cpp
@@ -800,7 +800,7 @@ Error FBXDocument::_parse_meshes(Ref<FBXState> p_state) {
 
 							// Enforce blend shape mask array format
 							for (int l = 0; l < Mesh::ARRAY_MAX; l++) {
-								if (!(Mesh::ARRAY_FORMAT_BLEND_SHAPE_MASK & (1 << l))) {
+								if (!(Mesh::ARRAY_FORMAT_BLEND_SHAPE_MASK & (static_cast<int64_t>(1) << l))) {
 									array_copy[l] = Variant();
 								}
 							}
